Added bézier path for independent enemy
Still needs some fixing though:
- Group leaving somehow broken
- Enemy never really reaches player
- Maybe get curve more curved?
- Check player-to-enemy collision

+++ Assignments/Megamaniac/megamaniac_go_data.h
@@ -41,7 +41,8 @@ typedef struct go_additional_data_enemy_indep_controller {
 	double control_y;
 	double target_x;
 	double target_y;
-	double t;
+	
+	int t;
 } go_additional_data_enemy_indep_controller_t;
 typedef go_additional_data_enemy_indep_controller_t* go_additional_data_enemy_indep_controller_p;
